feat(error): updating error handling  (#1723)
* fix(designsystem): Adding NoSSR-workarounds for designsystem flaws

* chore(deps): bump next and eslint to new major versions

* fix(tooltip): remove old tooltip, we are now using designsystemet

* fix(next 15): migrating to next 15 + new linting rules

* fix(search): create unique traveltags in the search

* feat(error): new way to handle errors

* feat(error): updating error handling

* feat(errorhandling): updating errorhandling

* Remove unused file

* fix(errorhandling): clean up

* fix(errorhandling): improving errorhandling

* fix

* chore(error): remove comments, unknown-errortype and columns-error-toast

* chore(error): remove unused return-expression in org-footer

* chore(error): move variable cheks into action-function for all settings, and move CountiesSelect out of MemberAdministration

---------

+++ tavla/app/(admin)/organizations/components/Footer/actions.ts
@@ -1,22 +1,34 @@
 'use server'
 import { isEmptyOrSpaces } from 'app/(admin)/edit/utils'
+import { getFormFeedbackForError } from 'app/(admin)/utils'
 import { userCanEditOrganization } from 'app/(admin)/utils/firebase'
+import { handleError } from 'app/(admin)/utils/handleError'
 import { firestore } from 'firebase-admin'
 import { revalidatePath } from 'next/cache'
 import { redirect } from 'next/navigation'
 import { TOrganizationID } from 'types/settings'
 
-export async function setFooter(oid: TOrganizationID, footer?: string) {
+export async function setFooter(
+    oid: TOrganizationID | undefined,
+    data: FormData,
+) {
+    if (!oid) return getFormFeedbackForError()
     const access = await userCanEditOrganization(oid)
     if (!access) return redirect('/')
 
-    await firestore()
-        .collection('organizations')
-        .doc(oid)
-        .update({
-            footer: !isEmptyOrSpaces(footer)
-                ? footer
-                : firestore.FieldValue.delete(),
-        })
-    revalidatePath(`organizations/${oid}`)
+    const message = data.get('footer') as string
+
+    try {
+        await firestore()
+            .collection('organizations')
+            .doc(oid)
+            .update({
+                footer: !isEmptyOrSpaces(message)
+                    ? message
+                    : firestore.FieldValue.delete(),
+            })
+        revalidatePath(`organizations/${oid}`)
+    } catch (e) {
+        return handleError(e)
+    }
 }
